refactor: change user to custumer in top layers
