+++ restaurante-aplicacao-banco-dados/src/main/java/br/com/nomeempresa/restaurante/adapters/inbound/entity/CustumerEntity.java
